+++ R/GPR_class.R
@@ -80,9 +80,10 @@ GPR_class <- nn_module(
   },
 
   # Unnormalised log likelihood for Gaussian Process
-  ldnorm = function(K, sigma2, beta) {
+  ldnorm = function(K, L, sigma2, beta) {
     log_lik <- .shrinkGPR_internal$jit_funcs$ldnorm(
       K = K,
+      L = L,
       sigma2 = sigma2,
       y = self$y,
       x_mean = self$x_mean,
@@ -151,9 +152,31 @@ GPR_class <- nn_module(
     return(list(zk = zk, log_det_J = log_det_J))
   },
 
+  # Function to generate data from t distribution
+  rt_torch = function(..., nu, device = "cpu", dtype = torch_float()) {
+    sample_shape <- c(...)
+
+    nu_t <- torch_scalar_tensor(nu, device = device, dtype = dtype)  # scalar
+    z <- torch_randn(!!!sample_shape, device = device, dtype = dtype)
+
+    g <- distr_gamma(
+      concentration = nu_t / 2,
+      rate = torch_scalar_tensor(0.5, device = device, dtype = dtype) # scalar
+    )
+    u <- g$sample(sample_shape = sample_shape)
+
+    # drop trailing singleton if present
+    if (u$ndim == z$ndim + 1 && u$size(u$ndim) == 1) {
+      u <- u$squeeze(-1)
+    }
+
+    z / torch_sqrt(u / nu_t)
+  },
+
   gen_batch = function(n_latent) {
     # Generate a batch of samples from the model
-    z <- torch_randn(n_latent, self$d, device = self$device)
+    # z <- torch_randn(n_latent, self$d, device = self$device)
+    z <- self$rt_torch(n_latent, self$d, nu = 2.1, device = self$device)
     return(z)
   },
 
@@ -169,6 +192,12 @@ GPR_class <- nn_module(
     sigma_zk <- zk_pos[, (self$x$shape[2] + 1)]
     lam_zk <- zk_pos[, (self$x$shape[2] + 2)]
 
+    # Res protector to avoid issues with 0 values
+    l2_zk <- res_protector_autograd(l2_zk)
+    lam_zk <- res_protector_autograd(lam_zk, tol = 1e-4)
+    sigma_zk <- res_protector_autograd(sigma_zk)
+
+
     if (!self$mean_zero) {
       beta <- zk_pos[, (self$x$shape[2] + 3):(self$x$shape[2] + 2 + self$x_mean$shape[2])]
       lam_mean <- zk_pos[, -1]
@@ -180,7 +209,18 @@ GPR_class <- nn_module(
     K <- self$kernel_func(l2_zk, lam_zk, self$x)
 
     # Calculate the components of the ELBO
-    likelihood <- self$ldnorm(K, sigma_zk, beta)$mean()
+    # This block uses robust chol if cholesky fails in ldnorm
+    tryCatch({
+      likelihood <- self$ldnorm(K, NULL, sigma_zk, beta)$mean()
+    }, error = function(ex) {
+      print(lam_zk)
+      single_eye <- torch_eye(self$N, device = self$device)
+      batch_sigma2 <- single_eye$`repeat`(c(sigma_zk$shape[1], 1, 1)) *
+        sigma_zk$unsqueeze(2)$unsqueeze(2)
+      L <- robust_chol(K + batch_sigma2, upper = FALSE)
+
+      likelihood <<- self$ldnorm(K, L, sigma_zk, beta)$mean()
+    })
 
     prior <- self$ltg(l2_zk, self$prior_a, self$prior_c, lam_zk)$sum(dim = 2)$mean() +
       self$ldf(lam_zk, 2*self$prior_c, 2*self$prior_a)$mean() +
@@ -212,11 +252,16 @@ GPR_class <- nn_module(
       # First, generate posterior draws by drawing random samples from the variational distribution.
       z <- self$gen_batch(nsamp)
       zk_pos <- self$forward(z)$zk
-      zk_pos <- res_protector_autograd(zk_pos)
 
       l2_zk <- zk_pos[, 1:self$x$shape[2]]
       sigma_zk <- zk_pos[, (self$x$shape[2] + 1)]
       lam_zk <- zk_pos[, (self$x$shape[2] + 2)]
+
+      # Res protector to avoid issues with 0 values
+      l2_zk <- res_protector_autograd(l2_zk)
+      lam_zk <- res_protector_autograd(lam_zk, tol = 1e-4)
+      sigma_zk <- res_protector_autograd(sigma_zk)
+
 
       if (!self$mean_zero) {
         beta <- zk_pos[, (self$x$shape[2] + 3):(self$x$shape[2] + 2 + self$x_mean$shape[2])]